+++ lsm-tree/src-core/Database/LSMTree/Internal/FS.hs
@@ -1,12 +1,13 @@
 module Database.LSMTree.Internal.FS (
     -- * Hard links
     hardLink
+  , hardLinkDirectoryRecursive
     -- * Copy file
   , copyFile
   ) where
 
 import           Control.ActionRegistry
-import           Control.Monad (void)
+import           Control.Monad (forM_, void)
 import           Control.Monad.Class.MonadThrow
 import           Control.Monad.Primitive (PrimMonad)
 
@@ -15,6 +16,7 @@ import           System.FS.API
 import qualified System.FS.API.Lazy as FSL
 import qualified System.FS.BlockIO.API as FS
 import           System.FS.BlockIO.API (HasBlockIO)
+import           Text.Printf (printf)
 
 {-------------------------------------------------------------------------------
   Hard links
@@ -43,6 +45,44 @@ hardLink hfs hbio reg sourcePath destinationPath = do
     withRollback_ reg
       (FS.createHardLink hbio sourcePath destinationPath)
       (FS.removeFile hfs destinationPath)
+
+{-# SPECIALISE
+  hardLinkDirectoryRecursive ::
+       HasFS IO h
+    -> HasBlockIO IO h
+    -> ActionRegistry IO
+    -> FS.FsPath
+    -> FS.FsPath
+    -> IO ()
+  #-}
+-- | Recursively create hard links for all the directory contents of the source
+-- path at the destination path.
+hardLinkDirectoryRecursive ::
+     (MonadMask m, PrimMonad m)
+  => HasFS m h
+  -> HasBlockIO m h
+  -> ActionRegistry m
+     -- | Source path
+  -> FS.FsPath
+     -- | Destination path
+  -> FS.FsPath
+  -> m ()
+hardLinkDirectoryRecursive hfs hbio reg sourcePath destinationPath = do
+    entries <- FS.listDirectory hfs sourcePath
+    forM_ entries $ \entry -> do
+      let sourcePath' = sourcePath FS.</> FS.mkFsPath [entry]
+          destinationPath' = destinationPath FS.</> FS.mkFsPath [entry]
+      isFile <- FS.doesFileExist hfs sourcePath'
+      if isFile then
+        hardLink hfs hbio reg sourcePath' destinationPath'
+      else do
+        isDirectory <- FS.doesDirectoryExist hfs sourcePath'
+        if isDirectory then do
+          hardLinkDirectoryRecursive hfs hbio reg sourcePath' destinationPath'
+        else
+          error $ printf
+            "hardLinkDirectoryRecursive: %s is not a file or directory"
+            (show sourcePath')
 
 {-------------------------------------------------------------------------------
   Copy file